+++ ExampleViewModel/ImageSearchTableViewCellModeling.swift
@@ -8,7 +8,7 @@
 import ReactiveCocoa
 
 public protocol ImageSearchTableViewCellModeling {
-    var id: UInt32 { get}
+    var id: UInt64 { get}
     var pageImageSizeText: String { get }
     var tagText: String { get }
